Rename _server_test file. Otherwise, `pub run test_runner` tries to run it and fails.
