+++ test/io_test.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 import 'package:test/test.dart';
 import 'package:oauth/oauth.dart' as oauth;
 import 'package:oauth/server_io.dart';
-import '_server_test.dart';
+import '_server_test_utils.dart';
 
 main() {
   HttpServer server;
